clean up commented out code
