--- internal/codeintel/uploads/internal/background/utils.go
@@ -1,103 +1,0 @@
-package background
-
-import (
-	"database/sql"
-	"time"
-
-	"github.com/keegancsmith/sqlf"
-	"github.com/lib/pq"
-
-	"github.com/sourcegraph/sourcegraph/internal/codeintel/shared/types"
-	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
-	dbworkerstore "github.com/sourcegraph/sourcegraph/internal/workerutil/dbworker/store"
-)
-
-// StalledUploadMaxAge is the maximum allowable duration between updating the state of an
-// upload as "processing" and locking the upload row during processing. An unlocked row that
-// is marked as processing likely indicates that the worker that dequeued the upload has died.
-// There should be a nearly-zero delay between these states during normal operation.
-const StalledUploadMaxAge = time.Second * 25
-
-// UploadMaxNumResets is the maximum number of times an upload can be reset. If an upload's
-// failed attempts counter reaches this threshold, it will be moved into "errored" rather than
-// "queued" on its next reset.
-const UploadMaxNumResets = 3
-
-var uploadWorkerStoreOptions = dbworkerstore.Options{
-	Name:              "codeintel_upload",
-	TableName:         "lsif_uploads",
-	ViewName:          "lsif_uploads_with_repository_name u",
-	ColumnExpressions: uploadColumnsWithNullRank,
-	Scan:              dbworkerstore.BuildWorkerScan(scanCompleteUpload),
-	OrderByExpression: sqlf.Sprintf(`
-		u.associated_index_id IS NULL DESC,
-		COALESCE(u.process_after, u.uploaded_at),
-		u.id
-	`),
-	StalledMaxAge: StalledUploadMaxAge,
-	MaxNumResets:  UploadMaxNumResets,
-}
-
-const visibleAtTipSubselectQuery = `SELECT 1 FROM lsif_uploads_visible_at_tip uvt WHERE uvt.repository_id = u.repository_id AND uvt.upload_id = u.id`
-
-var uploadColumnsWithNullRank = []*sqlf.Query{
-	sqlf.Sprintf("u.id"),
-	sqlf.Sprintf("u.commit"),
-	sqlf.Sprintf("u.root"),
-	sqlf.Sprintf("EXISTS (" + visibleAtTipSubselectQuery + ") AS visible_at_tip"),
-	sqlf.Sprintf("u.uploaded_at"),
-	sqlf.Sprintf("u.state"),
-	sqlf.Sprintf("u.failure_message"),
-	sqlf.Sprintf("u.started_at"),
-	sqlf.Sprintf("u.finished_at"),
-	sqlf.Sprintf("u.process_after"),
-	sqlf.Sprintf("u.num_resets"),
-	sqlf.Sprintf("u.num_failures"),
-	sqlf.Sprintf("u.repository_id"),
-	sqlf.Sprintf("u.repository_name"),
-	sqlf.Sprintf("u.indexer"),
-	sqlf.Sprintf("u.indexer_version"),
-	sqlf.Sprintf("u.num_parts"),
-	sqlf.Sprintf("u.uploaded_parts"),
-	sqlf.Sprintf("u.upload_size"),
-	sqlf.Sprintf("u.associated_index_id"),
-	sqlf.Sprintf("NULL"),
-	sqlf.Sprintf("u.uncompressed_size"),
-}
-
-func scanCompleteUpload(s dbutil.Scanner) (upload types.Upload, _ error) {
-	var rawUploadedParts []sql.NullInt32
-	if err := s.Scan(
-		&upload.ID,
-		&upload.Commit,
-		&upload.Root,
-		&upload.VisibleAtTip,
-		&upload.UploadedAt,
-		&upload.State,
-		&upload.FailureMessage,
-		&upload.StartedAt,
-		&upload.FinishedAt,
-		&upload.ProcessAfter,
-		&upload.NumResets,
-		&upload.NumFailures,
-		&upload.RepositoryID,
-		&upload.RepositoryName,
-		&upload.Indexer,
-		&dbutil.NullString{S: &upload.IndexerVersion},
-		&upload.NumParts,
-		pq.Array(&rawUploadedParts),
-		&upload.UploadSize,
-		&upload.AssociatedIndexID,
-		&upload.Rank,
-		&upload.UncompressedSize,
-	); err != nil {
-		return upload, err
-	}
-
-	upload.UploadedParts = make([]int, 0, len(rawUploadedParts))
-	for _, uploadedPart := range rawUploadedParts {
-		upload.UploadedParts = append(upload.UploadedParts, int(uploadedPart.Int32))
-	}
-
-	return upload, nil
-}
