+++ internal/codeintel/uploads/service_worker_handler.go
