+++ internal/codeintel/uploads/service_worker_handler_test.go
@@ -61,42 +61,12 @@ func TestHandle(t *testing.T) {
 	expectedCommitDateStr := expectedCommitDate.Format(time.RFC3339)
 	gitserverClient.CommitDateFunc.SetDefaultReturn("deadbeef", expectedCommitDate, true, nil)
 
-	// handler := &handler{
-	// 	dbStore:         mockDBStore,
-	// 	repoStore:       mockRepoStore,
-	// 	workerStore:     mockWorkerStore,
-	// 	lsifStore:       mockLSIFStore,
-	// 	uploadStore:     mockUploadStore,
-	// 	gitserverClient: gitserverClient,
-	// }
-
-	// svc := newService(
-	// 	mockDBStore,
-	// 	mockRepoStore,
-	// 	mockLSIFStore,
-	// 	gitserverClient,
-	// 	nil,
-	// 	nil,
-	// 	nil,
-	// 	nil,
-	// 	&observation.TestContext,
-	// )
-
-	// svc := &Service{
-	// 	store:           mockDBStore,
-	// 	repoStore:       mockRepoStore,
-	// 	workerutilStore: mockWorkerStore,
-	// 	lsifstore:       mockLSIFStore,
-	// 	policySvc:       nil,
-	// }
-
 	svc := &Service{
 		store:           mockDBStore,
 		repoStore:       mockRepoStore,
 		workerutilStore: mockWorkerStore,
 		lsifstore:       mockLSIFStore,
 		gitserverClient: gitserverClient,
-		// lsifstore:       mockLSIFStore,
 	}
 
 	requeued, err := svc.HandleRawUpload(context.Background(), logtest.Scoped(t), upload, mockUploadStore, observation.TestTraceLogger(logtest.Scoped(t)))
@@ -223,22 +193,6 @@ func TestHandleError(t *testing.T) {
 	// Set a different tip commit
 	mockDBStore.SetRepositoryAsDirtyFunc.SetDefaultReturn(errors.Errorf("uh-oh!"))
 
-	// handler := &handler{
-	// 	dbStore:         mockDBStore,
-	// 	repoStore:       mockRepoStore,
-	// 	workerStore:     mockWorkerStore,
-	// 	lsifStore:       mockLSIFStore,
-	// 	uploadStore:     mockUploadStore,
-	// 	gitserverClient: gitserverClient,
-	// }
-
-	// svc := &Service{
-	// 	store:           mockDBStore,
-	// 	repoStore:       mockRepoStore,
-	// 	workerutilStore: mockWorkerStore,
-	// 	lsifstore:       mockLSIFStore,
-	// }
-
 	svc := &Service{
 		store:           mockDBStore,
 		repoStore:       mockRepoStore,
@@ -291,20 +245,11 @@ func TestHandleCloneInProgress(t *testing.T) {
 		return "", &gitdomain.RepoNotExistError{Repo: repo.Name, CloneInProgress: true}
 	})
 
-	// handler := &handler{
-	// 	dbStore:         mockDBStore,
-	// 	repoStore:       mockRepoStore,
-	// 	workerStore:     mockWorkerStore,
-	// 	uploadStore:     mockUploadStore,
-	// 	gitserverClient: gitserverClient,
-	// }
-
 	svc := &Service{
 		store:           mockDBStore,
 		repoStore:       mockRepoStore,
 		workerutilStore: mockWorkerStore,
 		gitserverClient: gitserverClient,
-		// lsifstore:       mockLSIFStore,
 	}
 
 	requeued, err := svc.HandleRawUpload(context.Background(), logtest.Scoped(t), upload, mockUploadStore, observation.TestTraceLogger(logtest.Scoped(t)))
